run subprocess as seperate page
